requirements.txt
notion-client==2.2.1
python-dotenv==1.0.0
openai==1.12.0
reportlab==4.1.0 